requirements.txt
# Core dependencies
python-dotenv==1.0.0

# ML & NLP
torch==2.6.0
transformers==4.45.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2

# Document processing
pypdf==5.1.0
pdfplumber==0.11.4
beautifulsoup4==4.12.3
lxml==5.3.0

# Text processing
unidecode==1.3.8

# Web interface
streamlit==1.41.1
streamlit-chat==0.1.1

# CLI
typer==0.15.1
rich==13.9.4

# Utilities
pydantic==2.10.4
pyyaml==6.0.2
tqdm==4.67.1
requests==2.32.3

# Logging
loguru==0.7.3

# Data manipulation
numpy==1.26.4
pandas==2.2.3

# Development (opcional)
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
